Promote presets to CLI-wide logic
