+++ packages/app/src/cli/commands/app/info.ts
@@ -2,9 +2,9 @@ import {appFlags} from '../../flags.js'
 import {AppInterface} from '../../models/app/app.js'
 import {Format, info} from '../../services/info.js'
 import {load as loadApp} from '../../models/app/loader.js'
-import Command from '../../utilities/app-command.js'
 import {Flags} from '@oclif/core'
 import {output, path, cli} from '@shopify/cli-kit'
+import Command from '@shopify/cli-kit/node/base-command'
 
 export default class AppInfo extends Command {
   static description = 'Print basic information about your app and extensions'
